# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  wowrack-recruitment_internal_dto.UpdateStatusRequest:
    properties:
      duration_minutes:
        description: wajib jika status interview, misal 60
        type: integer
      interview_date:
        description: wajib jika status interview
        type: string
      notes:
        type: string
      status:
        enum:
        - screening
        - contacted
        - hr_interview
        - hiring_manager_interview
        - salary_negotiation
        - hired
        - rejected
        type: string
    required:
    - status
    type: object
host: localhost:8080
info:
  contact:
    email: dev@albyeicorp.com
    name: Albyei Corp Engineering
  description: API untuk sistem rekrutmen Wowrack/Albyei Corp. Mendukung multi-role
    (Admin, HR, Hiring Manager, Candidate).
  termsOfService: http://swagger.io/terms/
  title: Wowrack Recruitment API
  version: "1.0"
paths:
...
  /hr/active-vacancies:
    get:
      description: Get list of all active (published) job vacancies for HR dashboard
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Get active vacancies
      tags:
      - HR Applications
  /hr/active-vacancies/{id}:
    get:
      description: Get a single active vacancy detail by its ID (HR only)
      parameters:
      - description: Vacancy ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Get active vacancy by ID
      tags:
      - HR Applications
  /hr/applications:
    get:
      description: Get paginated list of all job applications (HR only)
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Get all applications
      tags:
      - HR Applications
  /hr/applications/{id}:
    get:
      description: Get a single application detail by its ID (HR only)
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Get application by ID
      tags:
      - HR Applications
  /hr/applications/{id}/status:
    patch:
      consumes:
      - application/json
      description: Change the status of a job application (HR only)
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: integer
      - description: Status update payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/wowrack-recruitment_internal_dto.UpdateStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Update application status
      tags:
      - HR Applications
...
  /hr/jobs/{id}/applications:
    get:
      description: Get paginated list of applications for a specific job (HR only)
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: integer
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Get applications by job
      tags:
      - HR Applications
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"